refactor: move get_cached_archive_for_link and Wheel to utils
