+++ tests/utils/test_cache.py
@@ -10,11 +10,14 @@ from unittest.mock import Mock
 import pytest
 
 from cachy import CacheManager
+from packaging.tags import Tag
 from poetry.core.packages.utils.link import Link
 
 from poetry.utils.cache import FileCache
 from poetry.utils.cache import get_cache_directory_for_link
+from poetry.utils.cache import get_cached_archive_for_link
 from poetry.utils.cache import get_cached_archives_for_link
+from poetry.utils.env import MockEnv
 
 
 if TYPE_CHECKING:
@@ -208,6 +211,105 @@ def test_get_cached_archives_for_link(mocker: MockerFixture) -> None:
 
     assert archives
     assert set(archives) == set(distributions.glob("demo-0.1.*"))
+
+
+@pytest.mark.parametrize(
+    ("link", "strict", "available_packages"),
+    [
+        (
+            "https://files.python-poetry.org/demo-0.1.0.tar.gz",
+            True,
+            [
+                Path("/cache/demo-0.1.0-py2.py3-none-any"),
+                Path("/cache/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl"),
+                Path("/cache/demo-0.1.0-cp37-cp37-macosx_10_15_x86_64.whl"),
+            ],
+        ),
+        (
+            "https://example.com/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl",
+            False,
+            [],
+        ),
+    ],
+)
+def test_get_not_found_cached_archive_for_link(
+    mocker: MockerFixture,
+    link: str,
+    strict: bool,
+    available_packages: list[Path],
+) -> None:
+    env = MockEnv(
+        version_info=(3, 8, 3),
+        marker_env={"interpreter_name": "cpython", "interpreter_version": "3.8.3"},
+        supported_tags=[
+            Tag("cp38", "cp38", "macosx_10_15_x86_64"),
+            Tag("py3", "none", "any"),
+        ],
+    )
+
+    mocker.patch(
+        "poetry.utils.cache.get_cached_archives_for_link",
+        return_value=available_packages,
+    )
+
+    archive = get_cached_archive_for_link(env, Path(), Link(link), strict=strict)
+
+    assert archive is None
+
+
+@pytest.mark.parametrize(
+    ("link", "cached", "strict"),
+    [
+        (
+            "https://files.python-poetry.org/demo-0.1.0.tar.gz",
+            "/cache/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl",
+            False,
+        ),
+        (
+            "https://example.com/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl",
+            "/cache/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl",
+            False,
+        ),
+        (
+            "https://files.python-poetry.org/demo-0.1.0.tar.gz",
+            "/cache/demo-0.1.0.tar.gz",
+            True,
+        ),
+        (
+            "https://example.com/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl",
+            "/cache/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl",
+            True,
+        ),
+    ],
+)
+def test_get_found_cached_archive_for_link(
+    mocker: MockerFixture,
+    link: str,
+    cached: str,
+    strict: bool,
+) -> None:
+    env = MockEnv(
+        version_info=(3, 8, 3),
+        marker_env={"interpreter_name": "cpython", "interpreter_version": "3.8.3"},
+        supported_tags=[
+            Tag("cp38", "cp38", "macosx_10_15_x86_64"),
+            Tag("py3", "none", "any"),
+        ],
+    )
+
+    mocker.patch(
+        "poetry.utils.cache.get_cached_archives_for_link",
+        return_value=[
+            Path("/cache/demo-0.1.0-py2.py3-none-any"),
+            Path("/cache/demo-0.1.0.tar.gz"),
+            Path("/cache/demo-0.1.0-cp38-cp38-macosx_10_15_x86_64.whl"),
+            Path("/cache/demo-0.1.0-cp37-cp37-macosx_10_15_x86_64.whl"),
+        ],
+    )
+
+    archive = get_cached_archive_for_link(env, Path(), Link(link), strict=strict)
+
+    assert Path(cached) == archive
 
 
 def test_get_cache_directory_for_link(tmp_path: Path) -> None:
